+++ app/app-text/components/scbd-branding/directives/xuser-notifications-panel.json
@@ -3,8 +3,8 @@
     "unread": "Unread",
     "read": "Read",
     "allTypes": "All Types",
-    "alerts": "Alerts",
-    "alert": "Alert",
+    "alerts": "Notifications",
+    "alert": "Notification",
     "messages": "Messages",
     "publishingRequests": "Publishing Requests",
     "publishingRequest": "Publishing Request",
@@ -13,8 +13,8 @@
     "recalledRequests": "Recalled Requests",
     "rejectedRequests": "Rejected Requests",
     "expiredRequests": "Expired Requests",
-    "searchByKeyword": "search by keyword",
-    "markAllRead": "mark all read",
+    "searchByKeyword": "Search by keyword",
+    "markAllRead": "Mark all read",
     "requestExpired": "Request Expired",
     "publishingRequestExpired": "Publishing request expired",
     "draftFolder": "The requested record has been placed back into the owners draft folder",
@@ -22,25 +22,25 @@
     "expired": "Publishing Requests are automatically expired by the system after 90 days",
     "ago": "ago",
     "deleteRequest": "Delete request",
-    "requestedBy": "Requested by:",
+    "requestedBy": "Requested by",
     "on": "on",
     "reminder": "Reminder",
     "reminderPendingApproval": "Reminder: Publishing request pending approval.",
     "reminderDeleteApproval": "Reminder: Delete request pending approval.",
     "requestRecalled": "Request recalled",
     "cancelled": "The requester has cancelled their request for publishing a record",
-    "recalledBy": "Recalled by:",
+    "recalledBy": "Recalled by",
     "requestRejected": "Request rejected",
     "publishingAuthority": "Publishing Authority",
     "bchNationalFocalPoint": "BCH National Focal Point",
     "hasRejectedForPublishing": "has rejected as request for publishing",
-    "rejectedBy": "Rejected by:",
-    "reason": "Reason :",
+    "rejectedBy": "Rejected by",
+    "reason": "Reason",
     "requestApproved": "Request approved",
     "hasApproved": "has approved a request and published the record",
-    "approvedBy": "Approved by:",
+    "approvedBy": "Approved by",
     "message": "Message",
-    "sentFrom": "Sent from:",
+    "sentFrom": "Sent from",
     "newRecordPublished": "A new record has been published",
     "noNotifications": "no notifications",
     "noNewRequests": "You have no new notifications or requests"
