+++ app/app-text/routes/common-routes-labels.json
@@ -28,7 +28,7 @@
   "reportedRecords":"Reported records",
   "reportCounts":"Report counts",
   "errorLogs":"Error logs",
-  "subscriptions":"Subscriptions",
+  "subscriptions":"Email Alerts",
   "userRoleReport":"User role report",
   "commonIssues":"Common issues",
   "matrix":"Matrix",
