+++ app/app-text/views/register/user-preferences/preferences.json
@@ -1,7 +1,3 @@
 {
-    "userPreferences": "User Preferences",
-    "emailNotification": "Email notifications and predefined searches",
-    "keepUpToDate": "Keep up to date when records are published.",   
-    "settingsTab": "Settings",
-    "alertsTab":"Alerts"
+    "alertsTab":"Email Alerts"
 }
