--- app/app-text/views/register/user-preferences/user-settings.json
@@ -1,13 +1,0 @@
-{
-    "settingsDescription" : "Your application settings",
-
-
-    "languageHeader":"Application Language",
-    "languageDescription":"Set your preferred language (login required)",
-    "languageDropdown":"Select language",
-    "frequencyHeader":"Email alerts frequency",
-    "frequencyDescription":"Set the frequency of email alerts you receive from the Clearing-House",
-    "frequencyDropdown":"Select frequency",
-
-    "updating": "updating"
-}
