+++ app/components/scbd-branding/directives/xuser-notifications-panel.html
@@ -142,7 +142,7 @@
                             <span> {{notification.data.documentInfo.title | lstring:locale }} </span>
                         </div>
                         <div class="d-none d-sm-inline-block notification-spacing5">
-                            <span class="notification-label">{{::('xuserNotificationsPanelT.requestedBy'|$translate)}} </span>
+                            <span class="notification-label">{{::('xuserNotificationsPanelT.requestedBy'|$translate)}}: </span>
                             <span> {{notification.createdBy_info.firstName}} {{notification.createdBy_info.lastName}} </span>
                             <span class="notification-label"> {{::('xuserNotificationsPanelT.on'|$translate)}} </span>
                             <span> {{notification.createdOn | formatDateWithTime}} </span>
@@ -180,7 +180,7 @@
                             <span> {{notification.data.documentInfo.title | lstring:locale }} </span>
                         </div>
                         <div class="d-none d-sm-inline-block notification-spacing5">
-                            <span class="notification-label">{{::('xuserNotificationsPanelT.requestedBy'|$translate)}} </span>
+                            <span class="notification-label">{{::('xuserNotificationsPanelT.requestedBy'|$translate)}}: </span>
                             <span> {{notification.createdBy_info.firstName}} {{notification.createdBy_info.lastName}} </span>
                             <span class="notification-label"> {{::('xuserNotificationsPanelT.on'|$translate)}} </span>
                             <span> {{notification.createdOn | formatDateWithTime}} </span>
@@ -219,7 +219,7 @@
                             <span> {{notification.data.documentInfo.title | lstring:locale }} </span>
                         </div>
                         <div class="d-none d-sm-inline-block notification-spacing5">
-                            <span class="notification-label">{{::('xuserNotificationsPanelT.recalledBy'|$translate)}} </span>
+                            <span class="notification-label">{{::('xuserNotificationsPanelT.recalledBy'|$translate)}}: </span>
                             <span> {{notification.createdBy_info.firstName}} {{notification.createdBy_info.lastName}} </span>
                             <span class="notification-label"> {{::('xuserNotificationsPanelT.on'|$translate)}} </span>
                             <span> {{notification.createdOn | formatDateWithTime}} </span>
@@ -262,7 +262,7 @@
                             <span> {{notification.data.documentInfo.title | lstring:locale }} </span>
                         </div>
                         <div class="d-none d-sm-inline-block notification-spacing5">
-                            <span class="notification-label">{{::('xuserNotificationsPanelT.rejectedBy'|$translate)}} </span>
+                            <span class="notification-label">{{::('xuserNotificationsPanelT.rejectedBy'|$translate)}}: </span>
                             <span> {{notification.createdBy_info.firstName}} {{notification.createdBy_info.lastName}} </span>
                             <span class="notification-label"> {{::('xuserNotificationsPanelT.on'|$translate)}} </span>
                             <span> {{notification.createdOn | formatDateWithTime}}</span>
@@ -310,7 +310,7 @@
                             <span> {{notification.data.documentInfo.title | lstring:locale }} </span>
                         </div>
                         <div class="d-none d-sm-inline-block notification-spacing5">
-                            <span class="notification-label">{{::('xuserNotificationsPanelT.approvedBy'|$translate)}} </span>
+                            <span class="notification-label">{{::('xuserNotificationsPanelT.approvedBy'|$translate)}}: </span>
                             <span> {{notification.createdBy_info.firstName}} {{notification.createdBy_info.lastName}} </span>
                             <span class="notification-label"> {{::('xuserNotificationsPanelT.on'|$translate)}} </span>
                             <span> {{notification.createdOn | formatDateWithTime}} </span>
