+++ app/views/register/notifications.html
@@ -5,7 +5,7 @@
     <div smooth-scroll duration="100" offset="200"></div>
     <span register-top-menu></span>
 
-    <div class=" innerPageContent register-content card-1-notrans m-0 mb-5 p-3">
+    <div class="innerPageContent register-content m-0 p-3">
 
         <div class="register-content-header bg-darkgrey color-white">
             <div class="p-2">
