+++ app/views/register/requests.html
@@ -4,7 +4,7 @@
 
     <span register-top-menu></span>
 
-    <div class="innerPageContent register-content card card-1-notrans m-0 mb-5 p-3">
+    <div class="innerPageContent register-content m-0 p-3">
 
         <div class="register-content-header bg-darkgrey color-white">
             <div class="p-2">
