+++ app/views/register/user-preferences/user-alerts.html
@@ -9,6 +9,33 @@
     </style>
     <div class="container-fluid">
         <div>
+            <div class="row m-0 p-2 bg-light mt-2">
+                <div class="col-7 col-md-8">                    
+                    <span class="bold">
+                        {{::('userAlertsT.frequencyHeader'|$translate)}}
+                    </span>
+                    <small>{{::('userAlertsT.frequencyDescription'|$translate)}}</small>
+                </div>
+                <div class="col-1 text-center color-darkgray">                   
+                </div>
+                <div class="col-1 text-center color-white">                   
+                </div>
+                <div class="col-1 text-center color-white">                   
+                </div>
+                <div class="col-1 offset-1 offset-md-0 text-center text-secondary">
+                    <div class="btn-group">
+                        <button class="btn btn-primary dropdown-toggle" type="button" data-bs-toggle="dropdown" aria-expanded="false" ng-disabled="updating">
+                            <span ng-if="!userSettings.alertFrequency">{{::('userAlertsT.frequencyDropdown'|$translate)}}</span>
+                            <span>{{frequencies[userSettings.alertFrequency]}}</span>
+                        </button>
+                        <ul class="dropdown-menu">
+                            <li ng-repeat="(key, frequency) in frequencies">
+                                <a class="dropdown-item" ng-class="{'active': userSettings.alertFrequency == key}" ng-click="updateFrequency(key)" href="#">{{frequency}}</a>
+                            </li>
+                        </ul>
+                      </div>
+                </div>
+            </div>
             <div class="row m-0 p-2 mt-2">
                 <h2 class="p-0">
                     <small class="fs-6">{{::('userAlertsT.systemAlertHeading'|$translate)}}</small>
