+++ app/views/register/user-preferences/user-alerts.js
@@ -7,6 +7,7 @@ import '~/services/main';
 import '~/views/register/directives/register-top-menu';
 import '~/components/scbd-angularjs-services/main';
 import userAlertsT from '~/app-text/views/register/user-preferences/user-alerts.json';
+import frequencies from '~/app-text/views/register/user-preferences/frequency.json'
 
     app.directive("userAlerts", ['$rootScope', 'ngDialog', '$routeParams', function ($rootScope, ngDialog, $routeParams) {
 
@@ -32,6 +33,8 @@ import userAlertsT from '~/app-text/views/register/user-preferences/user-alerts.
                     $scope.systemAlertsSubscription = [];
                     $scope.isABS = realm.is('ABS');
                     $scope.isDeleteAllow = false ;
+                    $scope.frequencies = frequencies;
+                    $scope.userSettings = {};
                     translationService.set('userAlertsT', userAlertsT); 
                     
                     const systemQueries = {
@@ -87,7 +90,7 @@ import userAlertsT from '~/app-text/views/register/user-preferences/user-alerts.
                     //==============================================================
                     $rootScope.$on('signIn', function (evt, user) {
                         $scope.user = user;
-                        loadSavedFilters();
+                        init();
                     });
 
                     //==============================================================
@@ -326,8 +329,67 @@ import userAlertsT from '~/app-text/views/register/user-preferences/user-alerts.
                             return name[locale]
                         else return name;
                     }
+                    
 
-                    loadSavedFilters();
+                    $scope.updateFrequency = async function(alertFrequency){
+                        $scope.userSettings.alertFrequency = alertFrequency;
+
+                        await updateUserSettings($scope.userSettings);
+                    };
+
+                    async function init(){
+                        $scope.updating = true;
+                        try{
+                            loadSavedFilters();
+                            const settings = await $http.get('/api/v2016/settings/'+`${realm.value}-${$scope.user.userID}`);
+                            
+                            if(settings)
+                                $scope.userSettings = settings.data;
+                        }
+                        catch(e){
+                            if(e.status != 404)
+                                safeApply(()=>$scope.error = e.data);
+                        }
+                        finally{
+                            safeApply(()=>$scope.updating = false);
+                        }
+                    }
+
+                    async function updateUserSettings(data){
+                        if(!data.realm)
+                            data.realm = realm.value;
+                        
+                        //backend schema was changed to consolidated all setting..to have ch based setting only, create realm + user userID
+                        if(!data.userId)
+                            data.userId = `${realm.value}-${$scope.user.userID}`;
+                                                    
+                        $scope.updating = true;
+                        $scope.error = undefined;
+
+                        try{
+                            let settingPromise;
+                            if(data._id)
+                                settingPromise = $http.put('/api/v2016/settings/'+data.userId, data);
+                            else
+                                settingPromise = $http.post('/api/v2016/settings', data);
+
+                            const settingResponse = await settingPromise;
+                            if(settingResponse.status == 200 && settingResponse.data.id)
+                                $scope.userSettings._id =  settingResponse.data.id;
+                        }
+                        catch(e){
+                            safeApply(()=>$scope.error = e.data);
+                        }
+                        finally{
+                            safeApply(()=>$scope.updating = false);
+                        }
+
+                    }
+                    function safeApply(fn) {
+                        ($scope.$$phase || $scope.$root.$$phase) ? fn() : $scope.$apply(fn);
+                    } 
+
+                    init();
                 }
             ]
         };
